+++ DatadogCrashReporting/Sources/ThirdPartyCrashReporter.swift
@@ -5,11 +5,14 @@
  */
 
 import Foundation
+import DatadogInternal
 
 /// An interface of 3rd party crash reporter used by the DatadogCrashReporting.
 internal protocol ThirdPartyCrashReporter {
     /// Initializes and enables the crash reporter.
     init() throws
+
+    // MARK: - Crash Reporting
 
     /// Tells if there is a crash report available.
     func hasPendingCrashReport() -> Bool
@@ -22,4 +25,8 @@ internal protocol ThirdPartyCrashReporter {
 
     /// Deletes the available crash report.
     func purgePendingCrashReport() throws
+
+    // MARK: - Backtrace Generation
+
+    func generateBacktrace(threadID: ThreadID) throws -> BacktraceReport
 }