+++ DatadogLogs/Tests/Log/LogEventBuilderTests.swift
@@ -25,6 +25,8 @@ class LogEventBuilderTests: XCTestCase {
         let randomOsName: String = .mockRandom()
         let randomOsVersion: String = .mockRandom()
         let randomOsBuildNumber: String = .mockRandom()
+        let randomName: String = .mockRandom()
+        let randomModel: String = .mockRandom()
         let randomArchitecture: String = .mockRandom()
 
         // Given
@@ -46,6 +48,8 @@ class LogEventBuilderTests: XCTestCase {
             context: .mockWith(
                 serverTimeOffset: 0,
                 device: .mockWith(
+                    name: randomName,
+                    model: randomModel,
                     osName: randomOsName,
                     osVersion: randomOsVersion,
                     osBuildNumber: randomOsBuildNumber,
@@ -61,11 +65,15 @@ class LogEventBuilderTests: XCTestCase {
             XCTAssertEqual(log.error?.kind, randomError.type)
             XCTAssertEqual(log.error?.message, randomError.message)
             XCTAssertEqual(log.error?.stack, randomError.stack)
+            XCTAssertEqual(log.error?.sourceType, "ios")
             XCTAssertEqual(log.attributes, randomAttributes)
             XCTAssertEqual(log.tags.map { Set($0) }, randomTags)
             XCTAssertEqual(log.serviceName, randomService)
             XCTAssertEqual(log.loggerName, randomLoggerName)
             XCTAssertEqual(log.threadName, randomThreadName)
+            XCTAssertEqual(log.dd.device.brand, "Apple")
+            XCTAssertEqual(log.dd.device.name, randomName)
+            XCTAssertEqual(log.dd.device.model, randomModel)
             XCTAssertEqual(log.dd.device.architecture, randomArchitecture)
             XCTAssertEqual(log.os.name, randomOsName)
             XCTAssertEqual(log.os.version, randomOsVersion)
@@ -89,6 +97,8 @@ class LogEventBuilderTests: XCTestCase {
         let randomNetworkInfo: NetworkConnectionInfo = .mockRandom()
         let randomCarrierInfo: CarrierInfo = .mockRandom()
         let randomServerOffset: TimeInterval = .mockRandom(min: -10, max: 10)
+        let randomName: String = .mockRandom()
+        let randomModel: String = .mockRandom()
         let randomOSVersion: String = .mockRandom()
         let randomOSBuild: String = .mockRandom()
 
@@ -99,6 +109,8 @@ class LogEventBuilderTests: XCTestCase {
             sdkVersion: randomSDKVersion,
             serverTimeOffset: randomServerOffset,
             device: .mockWith(
+                name: randomName,
+                model: randomModel,
                 osVersion: randomOSVersion,
                 osBuildNumber: randomOSBuild
             ),
@@ -139,6 +151,9 @@ class LogEventBuilderTests: XCTestCase {
             DDAssertDictionariesEqual(log.userInfo.extraInfo, randomUserInfo.extraInfo)
             XCTAssertEqual(log.networkConnectionInfo, randomNetworkInfo)
             XCTAssertEqual(log.mobileCarrierInfo, randomCarrierInfo)
+            XCTAssertEqual(log.dd.device.brand, "Apple")
+            XCTAssertEqual(log.dd.device.name, randomName)
+            XCTAssertEqual(log.dd.device.model, randomModel)
             XCTAssertEqual(log.os.version, randomOSVersion)
             XCTAssertEqual(log.os.build, randomOSBuild)
             expectation.fulfill()